refactor reindex script
